Move TestBuild into config_test.go, and all rotation tests into rotation_test.go
